Cleanup: Split out the WASAPI code into its own file.
