+++ Windows/main.cpp
@@ -39,7 +39,7 @@
 #include "Core/ConfigValues.h"
 #include "Core/SaveState.h"
 #include "Windows/EmuThread.h"
-#include "Windows/DSoundStream.h"
+#include "Windows/WindowsAudio.h"
 #include "ext/disarm.h"
 
 #include "Common/LogManager.h"
